Updates
Reorganize the project structure

Remove py extension from worker

Fix mistake so that "." and ".." paths work in worker

Add comment why we do that last pass through the folders. This is in case
new folders are added while we are running.

Some updates to the PDF

Make generation code simpler in gen.R